requirements.txt
pyautogen>=0.2,<0.3
openai>=1.0.0
ollama>=0.4.0
pandas>=2.0.0
numpy>=1.25.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
pydantic>=2.7.0
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openpyxl>=3.1.0
tabulate